refactor : entity조회를 dto 조회로 수정
