+++ api/src/main/java/org/meotppo/webti/service/propensityanalysis/PropensityAnalysisService.java
@@ -3,16 +3,14 @@ package org.meotppo.webti.service.propensityanalysis;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import org.meotppo.webti.domain.entity.jpa.developerprofile.WebDeveloperProfile;
 import org.meotppo.webti.domain.entity.jpa.question.Option;
 import org.meotppo.webti.domain.entity.jpa.question.Question;
 import org.meotppo.webti.domain.entity.type.MbtiType;
 import org.meotppo.webti.domain.repository.jpa.developertype.WebDeveloperProfileRepository;
 import org.meotppo.webti.domain.repository.jpa.question.QuestionRepository;
-import org.meotppo.webti.dto.file.ImageDto;
 import org.meotppo.webti.dto.propensityanalysis.PropensityAnalysisDto;
 import org.meotppo.webti.dto.propensityanalysis.PropensityOptionDto;
-import org.meotppo.webti.dto.propensityanalysis.PropensityProfileResponseDto;
+import org.meotppo.webti.domain.dto.propensityanalysis.PropensityProfileResponseDto;
 import org.meotppo.webti.dto.propensityanalysis.PropensityQuestionDto;
 import org.meotppo.webti.response.exception.common.WebDeveloperProfileNotFoundException;
 import org.springframework.stereotype.Service;
@@ -35,21 +33,10 @@ public class PropensityAnalysisService {
         type.append(propensityAnalysisDto.getTHINKING() > propensityAnalysisDto.getFEELING() ? "T" : "F");
         type.append(propensityAnalysisDto.getJUDGING() > propensityAnalysisDto.getPERCEIVING() ? "J" : "P");
 
-        WebDeveloperProfile developerProfile = webDeveloperProfileRepository.findByMbtiType(MbtiType.valueOf(type.toString()))
+        PropensityProfileResponseDto developerProfile = webDeveloperProfileRepository.findProfileByMbtiType(MbtiType.valueOf(type.toString()))
                 .orElseThrow(() -> new WebDeveloperProfileNotFoundException());
 
-        ImageDto imageDto = ImageDto.builder()
-                .url(developerProfile.getImage().getUrl())
-                .build();
-        
-        PropensityProfileResponseDto responseDto = PropensityProfileResponseDto.builder()
-                .result(developerProfile.getResult())
-                .description(developerProfile.getDescription())
-                .mbtiType(developerProfile.getMbtiType())
-                .imageDto(imageDto)
-                .build();
-
-        return responseDto;
+        return developerProfile;
     }
 
     public List<PropensityQuestionDto> getPropensityQuestions() {
