+++ domain/src/main/java/org/meotppo/webti/domain/dto/propensityanalysis/PropensityProfileResponseDto.java
@@ -1,12 +1,12 @@
-package org.meotppo.webti.dto.propensityanalysis;
+package org.meotppo.webti.domain.dto.propensityanalysis;
 
+import lombok.Data;
+import org.meotppo.webti.domain.dto.file.ImageDto;
 import org.meotppo.webti.domain.entity.type.MbtiType;
-import org.meotppo.webti.dto.file.ImageDto;
 
 import lombok.Builder;
-import lombok.Getter;
 
-@Getter
+@Data
 public class PropensityProfileResponseDto {
     private String result;
     private String description;
@@ -14,10 +14,10 @@ public class PropensityProfileResponseDto {
     private ImageDto imageDto;
 
     @Builder
-    public PropensityProfileResponseDto(String result, String description, MbtiType mbtiType, ImageDto imageDto) {
+    public PropensityProfileResponseDto(String result, String description, MbtiType mbtiType, String imageUrl) {
         this.result = result;
         this.description = description;
         this.mbtiType = mbtiType;
-        this.imageDto = imageDto;
+        this.imageDto = ImageDto.builder().url(imageUrl).build();
     }
 }
